requirements.txt
pyyaml~=6.0
commmons~=0.0.16
htmllistparse~=0.6.0
requests~=2.26.0
flask
